# Shizou.WebApi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Shizou.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/Account/Login": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "ShizouHttpClient_AccountLogin",
        "requestBody": {
          "description": "Password",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "description": "Sets the Identity cookie",
                "schema": {
                  "type": "string",
                  "description": "Sets the Identity cookie",
                  "format": ""
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/Account/SetPassword": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "ShizouHttpClient_AccountChangePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PasswordModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "description": "Sets the Identity cookie",
                "schema": {
                  "type": "string",
                  "description": "Sets the Identity cookie",
                  "format": ""
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/AniDbFiles/{id}/MarkWatched": {
      "put": {
        "tags": [
          "AniDbFiles"
        ],
        "operationId": "ShizouHttpClient_AniDbFilesMarkWatched",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/AniDbFiles/{id}/MarkUnwatched": {
      "put": {
        "tags": [
          "AniDbFiles"
        ],
        "operationId": "ShizouHttpClient_AniDbFilesMarkUnwatched",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/FileServer/{ed2K}": {
      "get": {
        "tags": [
          "FileServer"
        ],
        "operationId": "ShizouHttpClient_FileServerGet",
        "parameters": [
          {
            "name": "ed2K",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "416": {
            "description": "Range Not Satisfiable"
          },
          "206": {
            "description": "Partial Content",
            "content": {
              "application/octet-stream": { }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": { }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/FileServer/{ed2K}/Playlist": {
      "get": {
        "tags": [
          "FileServer"
        ],
        "operationId": "ShizouHttpClient_FileServerGetPlaylist",
        "parameters": [
          {
            "name": "ed2K",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "single",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/x-mpegurl": { }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/FileServer/{ed2k}/Fonts/{fontName}": {
      "get": {
        "tags": [
          "FileServer"
        ],
        "operationId": "ShizouHttpClient_FileServerGetFont",
        "parameters": [
          {
            "name": "ed2K",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fontName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "font/ttf": { },
              "font/otf": { }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/FileServer/{ed2K}/Subtitles/{index}": {
      "get": {
        "tags": [
          "FileServer"
        ],
        "operationId": "ShizouHttpClient_FileServerGetSubtitle",
        "parameters": [
          {
            "name": "ed2K",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/x-ssa": { }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PasswordModel": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}